+++ NEWS.md
@@ -1,7 +1,9 @@
 # splattr 0.0.0.9000 (development version)
 
+* Unexported `input_cvd()` until it can be made useful. (#12)
+
 * Added functions `hcl_dataframe()` and `hcl_table()`: work with colors in terms of 
-  hue, chroma, and luminance (#9).
+  hue, chroma, and luminance. (#9)
 
 * Added functions `ramp()` and `swatch()`: HTMLWidgets to display discrete and 
   contintuous palettes, respectively.