+++ NEWS.md
@@ -1,6 +1,6 @@
 # splattr 0.0.0.9000 (development version)
 
-* Unexported `input_cvd()` until it can be made useful. (#12)
+* Unexported `input_cvd()`; plan to restore when it can be made useful. (#12)
 
 * Added functions `hcl_dataframe()` and `hcl_table()`: work with colors in terms of 
   hue, chroma, and luminance. (#9)