+++ NEWS.md
@@ -1,5 +1,9 @@
-# splattr 0.0.0.9000 (development version)
+# splattr 0.0.0 (development version)
 
+* A (#16)
+
+* Added arguments `background` and `margin` to `ramp()` and `swatch()`, can help to simulate, for example, dark-mode. (#1, #2)
+ 
 * Unexported `input_cvd()`; plan to restore when it can be made useful. (#12)
 
 * Added functions `hcl_dataframe()` and `hcl_table()`: work with colors in terms of